test: fix broken end-to-end tests
We were using a Chevrolet Bolt in all the end-to-end tests, this caused
problems because the Bolt is not compatible with start / stop charge.
The tests previously worked because the Bolt was incorrectly marked as
compatible for start / stop charge for test mode.

This commit parameterizes the `runAuthFlow` and `getAuthClientParams`
methods so that we can run the end-to-end tests across multiple cars.